split TileCoord.tileCover into separate file
