+++ js/source/tile_coord.js
@@ -95,87 +95,3 @@ TileCoord.prototype.children = function() {
         new TileCoord(z, x + 1, y + 1, this.w, this.sourceMaxZoom)
     ];
 };
-
-// Taken from polymaps src/Layer.js
-// https://github.com/simplegeo/polymaps/blob/master/src/Layer.js#L333-L383
-
-function edge(a, b) {
-    if (a.row > b.row) { var t = a; a = b; b = t; }
-    return {
-        x0: a.column,
-        y0: a.row,
-        x1: b.column,
-        y1: b.row,
-        dx: b.column - a.column,
-        dy: b.row - a.row
-    };
-}
-
-function scanSpans(e0, e1, ymin, ymax, scanLine) {
-    var y0 = Math.max(ymin, Math.floor(e1.y0));
-    var y1 = Math.min(ymax, Math.ceil(e1.y1));
-
-    // sort edges by x-coordinate
-    if ((e0.x0 === e1.x0 && e0.y0 === e1.y0) ?
-            (e0.x0 + e1.dy / e0.dy * e0.dx < e1.x1) :
-            (e0.x1 - e1.dy / e0.dy * e0.dx < e1.x0)) {
-        var t = e0; e0 = e1; e1 = t;
-    }
-
-    // scan lines!
-    var m0 = e0.dx / e0.dy;
-    var m1 = e1.dx / e1.dy;
-    var d0 = e0.dx > 0; // use y + 1 to compute x0
-    var d1 = e1.dx < 0; // use y + 1 to compute x1
-    for (var y = y0; y < y1; y++) {
-        var x0 = m0 * Math.max(0, Math.min(e0.dy, y + d0 - e0.y0)) + e0.x0;
-        var x1 = m1 * Math.max(0, Math.min(e1.dy, y + d1 - e1.y0)) + e1.x0;
-        scanLine(Math.floor(x1), Math.ceil(x0), y);
-    }
-}
-
-function scanTriangle(a, b, c, ymin, ymax, scanLine) {
-    var ab = edge(a, b),
-        bc = edge(b, c),
-        ca = edge(c, a);
-
-    var t;
-
-    // sort edges by y-length
-    if (ab.dy > bc.dy) { t = ab; ab = bc; bc = t; }
-    if (ab.dy > ca.dy) { t = ab; ab = ca; ca = t; }
-    if (bc.dy > ca.dy) { t = bc; bc = ca; ca = t; }
-
-    // scan span! scan span!
-    if (ab.dy) scanSpans(ca, ab, ymin, ymax, scanLine);
-    if (bc.dy) scanSpans(ca, bc, ymin, ymax, scanLine);
-}
-
-TileCoord.cover = function(zoom, bounds, sourceMaxZoom, loadOverscaled) {
-    var z = Math.min(zoom, sourceMaxZoom);
-    var styleZ = loadOverscaled ? zoom : z;
-    var tiles = 1 << z;
-    var t = {};
-
-    function scanLine(x0, x1, y) {
-        var x, wx, coord;
-        if (y >= 0 && y <= tiles) {
-            for (x = x0; x < x1; x++) {
-                wx = (x % tiles + tiles) % tiles;
-                coord = new TileCoord(styleZ, wx, y, Math.floor(x / tiles), sourceMaxZoom);
-                t[coord.id] = coord;
-            }
-        }
-    }
-
-    // Divide the screen up in two triangles and scan each of them:
-    // +---/
-    // | / |
-    // /---+
-    scanTriangle(bounds[0], bounds[1], bounds[2], 0, tiles, scanLine);
-    scanTriangle(bounds[2], bounds[3], bounds[0], 0, tiles, scanLine);
-
-    return Object.keys(t).map(function(id) {
-        return t[id];
-    });
-};
